+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.0.2]
+
+- added a static method that initalize the JS class
+- all methods inside a JS class "UserList" are now private
+
 ## [1.0.1]
 
 - README.md update